Prompt for input before the slow part
This script takes a while to run, and then prompts for user input when it gets to setup_jira_auth. If you don't respond in time, it fails. If you try to rerun it, it gets confused because the git clones it created already exist.

Instead if we put setup_jira_auth very early in the script, you're more likely to see the interactive part immediately. Then you can leave the rest of the script to run in the background.