Remove setup_jira_auth (#69)
Having this step near the end of the script was a problem because people would often run the script unattended, and then this step would prompt for input, timeout, and fail. This prevented the final step of setting up gdb from running.

I think this Jira auth was originally used with rietveld: you run a script locally to both push a CR to rietveld and update the corresponding Jira ticket. Now instead we use Github CR and use a Github bot to update Jira.